Seventeenth day

+++ app/Http/Controllers/Admin/ProductController.php
@@ -41,9 +41,11 @@ class ProductController extends Controller
      */
     public function store(ProductRequest $request)
     {
-        $path = $request->file('image')->store('products');
         $params = $request->all();
-        $params['image'] = $path;
+        unset($params['image']);
+        if ($request->has('image')) {
+            $params['image'] = $request->file('image')->store('products');
+        }
 
         Product::create($params);
         return redirect()->route('products.index');
@@ -81,10 +83,20 @@ class ProductController extends Controller
      */
     public function update(ProductRequest $request, Product $product)
     {
-        Storage::delete($product->image);
-        $path = $request->file('image')->store('products');
         $params = $request->all();
-        $params['image'] = $path;
+        unset($params['image']);
+        if ($request->has('image')) {
+            Storage::delete($product->image);
+            $params['image'] = $request->file('image')->store('products');
+        }
+
+        foreach (['new', 'hit', 'recommend'] as $fieldName) {
+            if(!isset($params[$fieldName])){
+                $params[$fieldName] = 0;
+            }
+        }
+
+//        dd($params);
 
         $product->update($params);
         return redirect()->route('products.index');
